Move unmultiply_inplace to pixels crate
